add :: userUtil

+++ src/main/java/com/wandering/Do/domain/user/service/impl/DeleteApplyPromiseServiceImpl.java
@@ -2,9 +2,11 @@ package com.wandering.Do.domain.user.service.impl;
 
 import com.wandering.Do.domain.apply.entity.Application;
 import com.wandering.Do.domain.apply.repository.ApplicationRepository;
-import com.wandering.Do.domain.promise.entity.Promise;
 import com.wandering.Do.domain.promise.exception.PromiseNotFoundException;
+import com.wandering.Do.domain.promise.exception.UserNotMatchException;
+import com.wandering.Do.domain.user.entity.User;
 import com.wandering.Do.domain.user.service.DeleteApplyPromiseService;
+import com.wandering.Do.domain.user.util.UserUtil;
 import com.wandering.Do.global.annotation.ServiceWithTransactional;
 import lombok.RequiredArgsConstructor;
 
@@ -12,9 +14,15 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class DeleteApplyPromiseServiceImpl implements DeleteApplyPromiseService {
     private final ApplicationRepository applicationRepository;
+    private final UserUtil userUtil;
     public void execute(Long id) {
+        User user = userUtil.getCurrentUser();
+
         Application application = applicationRepository.findById(id)
                 .orElseThrow(PromiseNotFoundException::new);
+
+        if(!application.getUser().getId().equals(user.getId()))
+            throw new UserNotMatchException();
 
         applicationRepository.delete(application);
     }
